feat: Update PrivateGame.js file with the .jsx code and removed .jsx
